Helped Dastyni with a Problem
The problem intaled moving around in a circle formation around a pivoted
point. I did 1 Example which was a Square Box, which was not needed. So
this was the Second Attempt which Solved his Problem. I am uploading
both the Package containing the Help. As well as the Webplayer as
Example

Webplayer Link:
http://unitydevelopers.net/Examples/CircleAroundPivotPoint/